pdf-maker/requirements.txt
streamlit
reportlab
pillow
pyperclip
